+++ part1/courseinfo/my-app/package.json
