+++ part1/courseinfo/my-app/public/manifest.json
