+++ part1/courseinfo/my-app/src/components/Content/index.js
