+++ part1/courseinfo/my-app/src/components/Part/index.js
