+++ part1/courseinfo/my-app/src/components/Total/index.js
